Share message types between content and background
- Moves and consolidates the message types into the toplevel of the
  extension
