+++ projects/extension/src/background/types.ts
@@ -1,12 +1,4 @@
 import { AppMediator } from './AppMediator';
-// Messages that come from the app
-export type AppMessageType = 'associate' | 'rpc';
-
-export interface AppMessage {
-  type: AppMessageType;
-  payload: string; // smoldot name / json / message_id / subscription_id
-  subscription?: boolean;
-}
 
 export interface InitAppNameSpec {
   id: string,
@@ -15,26 +7,8 @@ export interface InitAppNameSpec {
   uAppName: string,
   chainSpec?: string
 }
-export interface Message extends MessageEvent {
-  data: {
-    error?: string;
-    id: string;
-    origin: string;
-    response?: string;
-    subscription?: string;
-    chainName?: string;
-    chainSpec?: string;
-  }
-}
 
-// Messages that we send to the app down through the port
-export type ExtensionMessageType = 'error' | 'rpc';
-
-export interface ExtensionMessage {
-  type: ExtensionMessageType;
-  payload: string;
-}
-
+// TODO: this is duplicated with `NetworkStatus` in ../types
 export type AppState = 'connected' | 'ready' | 'disconnecting' | 'disconnected';
 
 export interface MessageIDMapping {
